Fixing functions to work with the new order changes

+++ includes/class_order.php
@@ -107,9 +107,7 @@ class order {
 		$array = array(); 
 		if ($db->num_rows($result) > 0 ) {
 			$array = $db->fetch_array($result);
-			
-			//$addon->getAddonByBillingCycle();
-			
+						
 			$sql = "SELECT addon_id FROM  `<PRE>user_pack_addons` WHERE order_id = '{$id}'";
 			$result_addons = $db->query($sql);
 			$addon_list = array();
@@ -126,11 +124,16 @@ class order {
 	 * @return	array 
 	 * @author Julio Montoya <gugli100@gmail.com> BeezNest
 	 */	 
-	public function getAllOrdersToArray() {
+	public function getAllOrdersToArray($user_id = 0) {
 		global $main, $db, $style,$currency;
 		
-		// List invoices. :)
-		$result_order  = $db->query("SELECT * FROM `<PRE>user_packs` WHERE status <> '".ORDER_STATUS_DELETED."' ORDER BY id DESC");
+		if (empty($user_id)) {
+			$result_order  = $db->query("SELECT * FROM `<PRE>user_packs` WHERE status <> '".ORDER_STATUS_DELETED."' ORDER BY id DESC");	
+		} else {
+			$user_id = intval($user_id);
+			$result_order  = $db->query("SELECT * FROM `<PRE>user_packs` WHERE status <> '".ORDER_STATUS_DELETED."' AND userid = '".$user_id."' ORDER BY id DESC");
+		}
+		
 		
 		$query2 = $db->query("SELECT * FROM `<PRE>invoices` WHERE `is_paid` = 0 ");
 		$array2['list'] = "";
@@ -202,20 +205,20 @@ class order {
 
 			//Paid configuration links
 			$array['paid'] = ($array["is_paid"] == 1 ? "<span style='color:green'>Already Paid</span>" :
-														"<span style='color:red'>Unpaid</span>");														
-
+														"<span style='color:red'>Unpaid</span>");
 			$array['due'] =  ($array["is_paid"] == 1 ? '<span style="color:green">'.$array['due'].'</span>' :  '<span style="color:red">'.$array['due'].'</span>');
 			
 			$array['PACKAGE']		 = $package_name_list[$package_id];
 			$array['billing_cycle']  = $billing_cycle_name_list[$billing_cycle_id];
-			
-			$array['EDIT']  	= '<a href="index.php?page=orders&sub=edit&do='.$order_item['id'].'"><img src="../themes/icons/note_edit.png" title="Edit" alt="Edit" /></a>';			
-			$array['DELETE']  	= '<a href="index.php?page=orders&sub=delete&do='.$order_item['id'].'"><img src="../themes/icons/delete.png" title="Delete"  alt="Delete" /></a>';
-			$array['ADD_INVOICE']='<a href="index.php?page=orders&sub=add_invoice&do='.$order_item['id'].'"><img src="../themes/icons/note_add.png" title="Add invoice"  alt="Add invoice" /></a>';
-			
-			
-			//var_dump($array);
-			$array2['list'] .= $style->replaceVar("tpl/orders/list-item.tpl", $array);
+			if (empty($user_id)) {
+				$array['EDIT']  	= '<a href="index.php?page=orders&sub=edit&do='.$order_item['id'].'"><img src="../themes/icons/note_edit.png" title="Edit" alt="Edit" /></a>';			
+				$array['DELETE']  	= '<a href="index.php?page=orders&sub=delete&do='.$order_item['id'].'"><img src="../themes/icons/delete.png" title="Delete"  alt="Delete" /></a>';
+				$array['ADD_INVOICE']='<a href="index.php?page=orders&sub=add_invoice&do='.$order_item['id'].'"><img src="../themes/icons/note_add.png" title="Add invoice"  alt="Add invoice" /></a>';
+				$array2['list'] .= $style->replaceVar("tpl/orders/list-item.tpl", $array);
+			} else {
+				//This is for the client view
+				$array2['list'] .= $style->replaceVar("tpl/orders/list-item-client.tpl", $array);
+			}
 		}
 				
 		//$array2['num'] 			= mysql_num_rows($query);
@@ -248,7 +251,7 @@ class order {
 	 * @author Julio Montoya <gugli100@gmail.com> BeezNest
 	 */
 	public function getOrder($order_id, $read_only = false, $show_price = true) {
-		global $main, $db, $currency, $addon, $package;	
+		global $main, $db, $currency, $addon, $package, $billing, $user;	
 		
 		$order_info = $this->getOrderInfo($order_id);
 		
@@ -257,37 +260,28 @@ class order {
 		} else {			
 			$total = 0;			
 			$array['ID'] 		= $order_info['id'];
-			$user_id 			= $order_info['userid'];
+			$array['domain'] 	= $order_info['domain'];
+			$user_id 			= $order_info['userid'];			
+			$package_id 	  	= $order_info['pid'];
+			$billing_cycle_id 	= $order_info['billing_cycle_id'];
+			$addon_selected_list= $order_info['addons'];	
 			
 			//User info
-			$sql = "SELECT id, user, firstname, lastname FROM `<PRE>users` WHERE `id` = ".$user_id;
-			
-			$query_users 		= $db->query($sql);
-			$user_info  		= $db->fetch_array($query_users);
-			$array['USER'] 		=  $user_info['lastname'].', '.$user_info['firstname'].' ('.$user_info['user'].')';									
-			$array['CREATED_AT'] 	= date('Y-m-d', $order_info['signup']);			
-		
-			$addon_selected_list = $order_info['addons'];
-			
-			//Billing cycle
-			$sql = "SELECT id, pid, domain, billing_cycle_id FROM `<PRE>user_packs` WHERE `userid` = ".$user_id;
-			$query_domain 		= $db->query($sql);
-			
-			if($db->num_rows($query_domain) > 0) {					
-				$domain_info  		= $db->fetch_array($query_domain);		
+			$user_info = $user->getUserById($user_id);
 						
-				$array['domain'] 	= $domain_info['domain'];
-				$package_id 	  	= $domain_info['pid'];
-				$billing_cycle_id 	= $domain_info['billing_cycle_id'];					
+			$array['USER'] 		= $user_info['lastname'].', '.$user_info['firstname'].' ('.$user_info['user'].')';									
+			$array['CREATED_AT']= date('Y-m-d', $order_info['signup']);
+						
+			//Addon feature added			
+			if ($read_only) {
+				$show_checkboxes = false;								
+			} else {
+				$show_checkboxes = true;	
 			}
-			
-			//Addon feature added
-			
-			$addon_list = $addon->getAllAddonsByBillingCycleAndPackage($billing_cycle_id, $package_id);
-								
-			$array['ADDON'] = ' - ';
-			$addong_result_string = '';
-			$array['ADDON'] = 	$addon->generateAddonCheckboxesWithBilling($billing_cycle_id, $package_id, array_flip($addon_selected_list));
+			$result = $addon->showAllAddonsByBillingCycleAndPackage($billing_cycle_id, $package_id, array_flip($addon_selected_list), $show_checkboxes);
+			$array['ADDON'] = $result['html'];
+						 				
+			$total = $total + $result['total'];
 			
 			//Packages feature added				
 			$query = $db->query("SELECT * FROM `<PRE>packages`");
@@ -299,39 +293,35 @@ class order {
 			$package_list = array();		
 			while($package_data = $db->fetch_array($query)) {
 				$package_list[$package_data['id']] = array($package_data['name'], $package_data['id']);				
-			}
+			}			
+			$packages = $package->getAllPackagesByBillingCycle($billing_cycle_id);				
+			
+			$total = $total + $packages[$package_id]['amount'];
 			
 			if($read_only == true) {		
 				$array['PACKAGES'] = $package_list[$package_id][0];
-			} else {			
-				//$array['PACKAGES'] = $main->dropDown('package_id', $package_list, $package_id, 1 , '', array('onchange'=>'loadAddons(this);'));	
-				
-				$packages = $package->getAllPackagesByBillingCycle($billing_cycle_id);					
+				$array['PACKAGE_AMOUNT'] = $currency->toCurrency($packages[$package_id]['amount']);									
+			} else {		
 		   		$package_list = array();
 				foreach($packages as $package) {
-					$package_list[$package['id']] = array($package['name'].' - '.$currency->toCurrency($package['amount']), $package['id']);				
+					$package_list[$package['id']] = array($package['name'].' - '.$currency->toCurrency($package['amount']), $package['id']);									
 				}				
-				$array['PACKAGES'] = $main->dropDown('package_id', $package_list, $order_info['pid'], 1, '', array('onchange'=>'loadAddons(this);'));				
-			}
+				$array['PACKAGES'] = $main->dropDown('package_id', $package_list, $package_id, 1, '', array('onchange'=>'loadAddons(this);'));				
+			}		
 			
 			//Billing cycle
-			$query = $db->query("SELECT * FROM `<PRE>billing_cycles` WHERE status = ".BILLING_CYCLE_STATUS_ACTIVE);
-			if($db->num_rows($query) > 0) {						
-				$values = array();		
-				while($billing_cycle_data = $db->fetch_array($query)) {
-					$values[$billing_cycle_data['id']] = array($billing_cycle_data['name'], $billing_cycle_data['id']);				
-				}
-			}
-			$array['BILLING_CYCLES'] .= $main->dropDown('billing_cycle_id', $values, $billing_cycle_id, 1,'', array('onchange'=>'loadPackages(this);'));
-			
-			$user_pack_status = $main->getOrderStatusList();
-			
-			$new_order_list = array();
-			foreach($user_pack_status as $key=>$value) {
-				$new_order_list[] = array($value, $key);
-			}
-			
-			$array['STATUS'] = $main->dropDown('status', $new_order_list, $order_info['status']);	
+			$billing_list = $billing->getBillingCycles();
+			foreach($billing_list as $billing_item) {
+				$billing_list[$billing_item['id']] = $billing_item['name'];
+			}						
+			if ($read_only) {
+				$array['BILLING_CYCLES'] = $billing_list[$billing_cycle_id];
+			} else {
+				$array['BILLING_CYCLES'] = $main->createSelect('billing_cycle_id', $billing_list, $billing_cycle_id, 1,'', array('onchange'=>'loadPackages(this);'));
+			}									
+			$array['STATUS'] = $main->createSelect('status', $main->getOrderStatusList(), $order_info['status']);			
+			$array['TOTAL'] = $currency->toCurrency($total);
+			  	
 			return $array;
 		}
 	}
